Add further fixed point implementation code.
darcs-hash:20071016180556-d1b56-596b7d28cddfe85270e3d3592b82f8f322f7b8b6.gz

+++ d2/trans_single.h
@@ -170,7 +170,7 @@ private:
 			 * transformation.
 			 */ 
 			
-			ale_pos theta = eu[2] * M_PI / 180;
+			ale_pos theta = (double) eu[2] * M_PI / 180;
 
 			a = cos(theta) * scale_factor;
 			b = sin(theta) * scale_factor;
